Move factories tests to own file
